Refactor code for reuseability
