+++ client/src/hooks/use-artwork-data.ts
