+++ client/src/pages/artwork/[id].tsx
@@ -2,8 +2,9 @@ import { GetServerSideProps } from "next";
 import Image from "next/image";
 import { JSX } from "react";
 
-import ButtonGallery from "@/components/ui/goBack";
-import { generateMockArtwork } from "@/hooks/useArtworkData";
+import GoBackButton from "@/components/ui/go-back-button";
+import ImagePlaceholder from "@/components/ui/image-placeholder";
+import { generateMockArtwork } from "@/hooks/use-artwork-data";
 import api from "@/lib/api";
 import { Art } from "@/types/art";
 
@@ -109,7 +110,7 @@ export default function ArtworkPage({ artwork }: ArtworkPageProps) {
           data-layer="< Gallery"
           className="Gallery text-light-1 h-10 justify-start font-['DM_Sans'] text-3xl font-bold leading-10 tracking-tight"
         >
-          <ButtonGallery />
+          <GoBackButton url="/artwork" label="Gallery" />
         </div>
       </div>
       <div
@@ -126,26 +127,7 @@ export default function ArtworkPage({ artwork }: ArtworkPageProps) {
               className="relative block sm:h-auto sm:max-w-full md:max-h-full"
             />
           ) : (
-            // in case fail to load image or no image in db yet
-            <div
-              data-layer="Placeholder image"
-              className="PlaceholderImage bg-light-2 flex h-[500px] w-[500px] items-center justify-center rounded-[10px]"
-            >
-              <div data-svg-wrapper data-layer="Vector" className="Vector">
-                <svg
-                  width="96"
-                  height="96"
-                  viewBox="0 0 96 96"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M96 85.3333V10.6667C96 4.8 91.2 0 85.3333 0H10.6667C4.8 0 0 4.8 0 10.6667V85.3333C0 91.2 4.8 96 10.6667 96H85.3333C91.2 96 96 91.2 96 85.3333ZM29.3333 56L42.6667 72.0533L61.3333 48L85.3333 80H10.6667L29.3333 56Z"
-                    fill="var(--neutral-1, #1B1F4C)"
-                  />
-                </svg>
-              </div>
-            </div>
+            <ImagePlaceholder />
           )}
         </div>
         <div
